refactor: move three responsibilities out of at-baseline modules to satisfy the architecture gate
Codex review, PR #1184, third round -- three [debt-no-growth]/[new-test-size]
errors from `scripts/check_architecture.py`, all correct, all mine.

My own process failure, worth stating: the `architecture` step IS in
`verify.py --profile pr` and it PASSED on my first full run. I broke it in the
two follow-up commits, where I validated with `--only lint,typecheck,docs-build`
plus targeted pytest instead of re-running the profile. A narrower gate than
the one CI runs is not a gate.

Each is fixed by moving responsibility to a proper owner, never by raising a
baseline -- AGENTS.md is explicit that trimming a file to fit is the wrong
move and the entry should shrink because something left it.

1. `cli_compare_release_matrix.py` 1052 -> 996 (baseline 1025).
   `_resolve_release_package_side` moves to a new owner. Its siblings had no
   room either (`cli_compare_release_helpers.py` 1789/1790,
   `cli_compare_release_pairwise.py` 927/927), and the responsibility is real
   rather than a slice made to fit: Phase 7j gave it a second job beyond
   unpacking a stored package -- rendering the side-correct `--variant`
   remediation the engine deliberately cannot.

   First attempt put it at a flat `cli_*` root, which the gate rejected
   ([frozen-root-family]/[root-module]) and was right to: a front end
   translating an engine error into a usage error is exactly what ADR-061's
   `frontends` ring owns. It now sits at
   `frontends/cli/release_variant_operand.py`, beside `release_dry_run.py`/
   `release_summary.py`, as a deliberate import leaf so the consuming module
   gains no edge into the CLI-registration cycle. The move required
   re-rooting the function's lazy relative imports one package deeper.

2. `tests/regressions/manifest.py` 1403 -> 1210 (baseline 1347).
   Sharded on a real axis rather than by line count: bug classes about
   abicheck's own delivery surface (the CLI a user types, the CI that ships
   it, the developer tooling around both) move to
   `manifest_tool_surface.py`, leaving the analysis classes behind. A defect
   in that group is found by running the tool's own interfaces, never by
   comparing two libraries. `manifest.py` keeps the registry's behavior and
   becomes an assembly point over per-axis lists -- the same shape
   `abicheck/change_registry.py` already has over `model/change_catalog/`.
   All 25 classes still registered, ids still unique.

3. `tests/test_cli_compare_release_project_snapshot_package.py` 1250 -> 889
   (test maximum 1200). The variant tests move to
   `test_cli_compare_variant_selection.py`, split by contract axis, not size:
   the original module is about stored-versus-live release *parity*, the new
   one about the `--variant` *selector* -- its resolution primitive, its CLI
   surface, and whether its errors give advice that works. Fixture helpers
   are imported, not copied.

Verified with the full `verify.py --profile pr` this time: architecture 0
errors, 17/21 steps pass, coverage 96.39%. The three remaining failures
(fmt-check, ai-readiness, unit-pr's two `test_ai_readiness` cases) all
reproduce on unmodified origin/main -- the ADR-037/044/046 `**Verified:**`
receipts naming the unreachable commit 40f0063.
