+++ abicheck/cli_compare_release_matrix.py
@@ -75,6 +75,7 @@ from .frontends.cli.options.params import (
 from .frontends.cli.release_summary import (  # moved (ADR-065 S2), re-exported
     _write_release_summary_file as _write_release_summary_file,
 )
+from .frontends.cli.release_variant_operand import _resolve_release_package_side
 from .model import AbiSnapshot
 from .report.comparison_scope import ComparisonScopeTerms
 from .workflows.extraction import package_component_inventory
@@ -785,63 +786,6 @@ def _strip_diff_results_and_adjust_verdict(
     ) < _RELEASE_VERDICT_ORDER.get("COMPATIBLE_WITH_RISK", 0):
         worst_verdict = "COMPATIBLE_WITH_RISK"
     return worst_verdict
-
-
-def _resolve_release_package_side(
-    side_dir: Path,
-    variant_id: str | None,
-    make_temp_dir: Callable[[str], Path],
-    *,
-    side: str,
-) -> dict[str, Path] | None:
-    """``None`` when *side_dir* is not a stored `ProjectSnapshot` package
-    directory -- the caller falls back to its existing live-discovery path
-    unchanged. Otherwise, *side_dir* is unpacked via
-    `workflows.release_package.resolve_release_package_map` into
-    the same canonical-key -> `Path` shape `_build_match_map` builds from a
-    live directory of `.so` files (ADR-062 A1.7), so `_match_release_keys`'s
-    own ``set(old_map) & set(new_map)`` matches a stored-side library
-    against a live-side or another stored-side one by the identical key.
-
-    *side* is ``"old"`` or ``"new"`` -- the `--variant` prefix naming *this*
-    operand. It is used only to render an ambiguous-variant error's
-    remediation example, which is why it lives here and not in the engine:
-    see :class:`~abicheck.errors.AmbiguousVariantSelectionError`.
-    """
-    if not side_dir.is_dir():
-        return None
-    from .workflows.release_package import resolve_release_package_map
-    from .workflows.storage import is_project_snapshot_package_dir
-
-    if not is_project_snapshot_package_dir(side_dir):
-        return None
-    from .errors import AmbiguousVariantSelectionError, SnapshotError
-
-    dest_root = make_temp_dir("abicheck_relpkg_")
-    try:
-        return resolve_release_package_map(
-            side_dir, variant_id=variant_id, dest_root=dest_root
-        )
-    except AmbiguousVariantSelectionError as exc:
-        # The engine states the fact and carries the ids but names no flag,
-        # because only here is it known *which side* this package is --
-        # and a bare `--variant ID` would apply to both sides, so it is not
-        # safe advice when only one side is ambiguous (Codex review, PR
-        # #1184, second round). Append the side-correct example, and only
-        # when there is a real id to name: a package declaring zero
-        # variants has nothing to select.
-        example = (
-            f" (e.g. --variant {side}={exc.variant_ids[0]})" if exc.variant_ids else ""
-        )
-        raise click.UsageError(f"{exc}{example}") from exc
-    except (KeyError, ValueError, OSError, SnapshotError) as exc:
-        # Ambiguous variant, a same-key collision (ValueError), a missing/
-        # unreadable ref (OSError), an object absent from objects/ entirely
-        # (KeyError, DirectoryObjectStore.get's own error on a truncated
-        # package; CodeRabbit review), or a corrupt document (SnapshotError)
-        # are all usage errors, translated like `_build_match_map`'s own
-        # `AmbiguousLibraryMatchError` (Codex review).
-        raise click.UsageError(str(exc)) from exc
 
 
 def _prepare_compare_release_inputs(
